Refactor streaming video handler architecture for multi-model support
